requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
Django>=1.11.23
django-annoying==0.10.4
django-crispy-forms==1.7.2
django-filter==2.0.0
django-localflavor==2.2
django-registration-redux==2.6
django-rest-framework-docs==0.1.7
django-rest-multiple-models==2.1.3
django-rest-raml==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
enum34==1.1.6
gunicorn==19.9.0
httpie>=1.0.3
idna==2.8
ipaddress==1.0.22
itypes==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
jsonpickle==1.2
jsonref==0.2
markdown2==2.3.8
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2==2.8.3
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
python-stdnum==1.11
pytz==2019.1
PyYAML==5.1.2
raml-codec==0.1.4
ramlfications==0.1.9
requests==2.22.0
simplejson==3.16.0
six==1.12.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.25.3
xmltodict==0.12.0
